requirements.txt
faster-whisper
paho-mqtt
SpeechRecognition
pyaudio
scikit-learn
joblib
numpy
sounddevice
soundfile
edge-tts 
asyncio